Fix error in functional test 910
The current code for test 910 references self.keystone_sentry
this was replaced by an array in commit
4d2ab6668f8601a17bbda1bf89ab7633f22d8d3d.

There was probably a race between the commits and functional
tests was not run on a updated version of the tree prior to
commit 49f993988a1d99a1f31962dd3ba278a4647d8841.

After the addition of running keystone clustered in commit
4d2ab6668f8601a17bbda1bf89ab7633f22d8d3d the test in 910 has
transient failures. Add call to set_api_version after removal
and addition of percona-cluster relation as it will retry and
verify authentication accross all nodes making sure that
configuration has settled before performing final test.

Change-Id: Ib5505adb0ace7c86384bb922008b9a8d73f60f24