Unset DB init flag on shared-db relation removal
The keystone charm sets db-initialised to true after initializing the
database the first time.  However, if the database application is
removed, this flag is not unset.

This results in breakage on attempts to re-add a shared-db relation
with a new database application, as the charm will not attempt to
re-initialize the database prior to doing DB operations.

This fix simply ensures that we unset this flag prior to finalizing
removal of the shared-db relation.

Change-Id: I78ae12fda05ce006939b2d90a3d738bacb815915
Closes-Bug: #1797229