Minor updates to tox.ini for coverage target
To be in line with the other recent updates, changing the
charmhelpers/* exclusion to */charmhelpers/*, which more reliably
excludes the charmhelpers code from coverage tracking.  While this
presently does nothing differently for this charm, it's probably
better for this code to be in line with the other charms rather than
being different for no obvious reason.

Also excluding the unit_tests directory.

Change-Id: Idcc3bb494eb632b5f971ee6c44bca5284a7839a1