Fix manager.py launch python version to match installed payload
The manager.py helper is launched with either py2 or py3 depending on
the version of the payload.  However, it was using openstack-origin
rather than actually picking up the actual version of the installed
software.  This patch ensures that the manager.py is run with the
python version of the installed payload, which resolves a bug when
upgrading to rocky (which is the py2 to py3 transition of the payload).

Also fixes an os.execl() bug which changed behaviour from python2 to
python3.

Closes-bug: #1818988
(cherry picked from commit 627ab761dd4375bc1e5d5de3c703136495986d78)

Change-Id: Ie56374f5882619239f544aaeea4acf92d3d0879d