Ensure that manager.py exits cleanly if charm closes socket
In the related bug, the manager.py was crashing as it received
(essentially) a socket close which resulted in no data being received.
This wasn't handled properly resulting in a crash in the manager.py.
This change simply allows the manager.py to exit cleanly, and the charm
can restart it if it needs to.

Change-Id: I76a435f48cdb8c201b0ddc529da947e6ed9ec031
Related-Bug: #1822063