Backport Zaza reconfiguration
By splitting up Zaza from the openstack-charms specific tests,
we can more easily guarantee test stability, while not harming
out ability to improve on the testing library during a cycle.

This change backports the zaza namespace change landing, and
adds a stable branch of the zaza-openstack-charms project to
the test-dependencies.

Change-Id: I6916f6cbc96facb5d8277351c456a28c4c2d517e