Ensure that keystone service is paused if needed on series upgrade
During series upgrade, the keystone packages get re-installed as the
underlying Linux has been upgraded and new package sets are updated and
then pulled in.  For trusty->xenial this means that keystone.service
gets enabled which then breaks haproxy.  On install, on xenial+, the
keystone.service is disabled in the install hook.  This just replicates
this in the series-upgrade hook.

Change-Id: Ic5ed6cf354d5545b9e554e205a048955a381e0f5
Closed-Bug: #1839021