Send notification on identity-service relation.
Services can optionally request notifications of other services endpoint
changes. They do this by sending a space seperated list of service names
that they wish to be notified of down the identity-service relation e.g

    subscribe_ep_change="placement neutron"

If the endpoints change for any service in the list then a notification is
sent back with a nonce. e.g. if the neutron ep changes the charm will
recieve a json encoded dict of changes:

    'ep_changed': '{"neutron": "1c261658"}'

This removes the need for charms to have two relations for ep notification
changes and allows applications to specify which endpoints they are
interested in.

Closes-Bug: #1862974

Change-Id: I03667af35022c352ea1cf817d97a6a50c2fb5c5d