Add focal-ussuri and bionic-ussuri bundle
This patch adds a focal-ussuri and bionic-ussuri bundles to the tests
for the charm.  Note that this requires a modified zaza (to use the
force_deploy option).

The patchset includes a charm-helpers sync to bring in Py3.8 compatible
charm-helpers for focal support.

It also requires Py3.8 compatible versions of hacluster and glance
charms.  See:

 - https://review.opendev.org/#/c/711616/ - charm-glance
 - https://review.opendev.org/#/c/711617/ - charm-hacluster

Change-Id: I78569b8c51626177164084e4d1f5e18d40f4c3d8