Modifying wsgi template to add optional config for Kerberos authentication
For the keystone-kerberos charm to configure correctly
with the keystone charm, the apache config needs to be
edited. Also, the alias needs the info of the keystone
context for the location of the script, hence cannot
be put in a separate optional config. It does not affect
non kerberos configuration.

Change-Id: I2842b628f0c3f88bd7efd735f76926b5d34cc108