Silence the YAML NoneType warning for password_security-compliance
If the value is not set (the default) then the context causes a error
message to be emitted to the Juju log.  Whilst this doesn't affect the
operations of the charm, it is likely to cause support requests.  This
change only attempts to decode the config option if it is actually set.

Change-Id: I0ccccdab92cb46f7354732f9d620797910ba55c0